% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.access}
\alias{aue.access}
\title{EcoGenetics slot standard notation. Returns an accessor to the slot <ecoslot>
of the object <X>}
\usage{
aue.access(ecoslot, X)
}
\arguments{
\item{ecoslot}{Slot of EcoGenetics object}

\item{X}{object name}
}
\description{
EcoGenetics slot standard notation. Returns an accessor to the slot <ecoslot>
of the object <X>
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

