% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.check_class}
\alias{aue.check_class}
\title{Obtain the classes for each column of a data frame}
\usage{
aue.check_class(X)
}
\arguments{
\item{X}{factor}
}
\description{
Obtain the classes for each column of a data frame
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{export}
