% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.circle}
\alias{aue.circle}
\title{Circle perimeter}
\usage{
aue.circle(mat, r, x0, y0, smooth = 100)
}
\arguments{
\item{mat}{Input raster matrix.}

\item{r}{Radius of the circle in pixels.}

\item{x0}{Circle center- x position.}

\item{y0}{Circle center- y position.}

\item{smooth}{Smoothing factor.}
}
\description{
Circle perimeter
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}
