% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.circle.w}
\alias{aue.circle.w}
\title{Radial distance to a point.}
\usage{
aue.circle.w(mat, x0, y0)
}
\arguments{
\item{mat}{Input raster matrix.}

\item{x0}{Circle center- x position.}

\item{y0}{Circle center- y position.}
}
\description{
Radial distance to a point.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

