% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.dataAngle}
\alias{aue.dataAngle}
\title{Angles for an XY coordinates matrix}
\usage{
aue.dataAngle(XY, maxpi = FALSE, deg = FALSE, latlon = FALSE)
}
\arguments{
\item{XY}{XY matrix with projected coordinates}

\item{maxpi}{angles bounded between 0 - pi?}

\item{deg}{angles in decimal degrees?}

\item{latlon}{Are the coordinates in decimal degrees format? Defalut FALSE. If TRUE,
the coordinates must be in a matrix/data frame with the longitude in the first
column and latitude in the second. The position is projected onto a plane in
meters with the function \code{\link[SoDA]{geoXY}}.}
}
\description{
Angles for an XY coordinates matrix
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{export}
