% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.df2image}
\alias{aue.df2image}
\title{Transforming a data frame into a raster}
\usage{
aue.df2image(x, origin = c("upperleft", "lowerleft"))
}
\arguments{
\item{x}{output of aue.image2df}

\item{origin}{Origin of the reference for the coordinates. Default: upperleft.}
}
\description{
This is the inverse function of aue.image2df
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

