% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.diamond}
\alias{aue.diamond}
\title{Solid diamond}
\usage{
aue.diamond(mat, r, x0, y0)
}
\arguments{
\item{mat}{Input raster matrix.}

\item{r}{Radius of the square in pixels.}

\item{x0}{Square center- x position.}

\item{y0}{Square center- y position.}
}
\description{
Solid diamond
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

