% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.dummy2af}
\alias{aue.dummy2af}
\title{Converion from dummy allele matrix to frequencies}
\usage{
aue.dummy2af(df, loc_fac)
}
\arguments{
\item{df}{data frame with alleles coded in dummy format}

\item{loc_fac}{factor with the locus of each allele}
}
\description{
Converion from dummy allele matrix to frequencies
}
