% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.ellipse}
\alias{aue.ellipse}
\title{Solid ellipse}
\usage{
aue.ellipse(mat, a = 1, b = 1, x0, y0, theta = 0)
}
\arguments{
\item{mat}{Input raster matrix.}

\item{a}{x radius (semimajor axis)}

\item{b}{y radius (seminimor axis)}

\item{x0}{Ellipse center- x position.}

\item{y0}{Ellipse center- y position.}

\item{theta}{rotation angle in degrees.}

\item{r}{Ellipse expansion factor. Default = 1}
}
\description{
Solid ellipse
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{internal}
