% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.formatLine}
\alias{aue.formatLine}
\title{Remove spaces in a line of text}
\usage{
aue.formatLine(X)
}
\arguments{
\item{X}{character string}
}
\description{
Remove spaces in a line of text
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}
