% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.fqal}
\alias{aue.fqal}
\title{Allelic frequencies}
\usage{
aue.fqal(eco, grp = NULL)
}
\arguments{
\item{eco}{Object of class ecogen.}

\item{grp}{Column in the slot S for summary by group.}
}
\description{
Allelic frequencies
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

