% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.genlab}
\alias{aue.genlab}
\title{Generation of generic labels of constant length}
\usage{
aue.genlab(base, n)
}
\arguments{
\item{base}{a character string}

\item{n}{number of labels}
}
\description{
Generation of generic labels of constant length
}
\author{
Thibaut Jombart
}
\keyword{internal}

