% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.geom.dist}
\alias{aue.geom.dist}
\title{Weight matrices based on different geometries}
\usage{
aue.geom.dist(
  XY,
  geometry = c("ellipse", "square"),
  d = 1,
  a = 1,
  b = 1,
  row.sd = FALSE,
  theta = 0
)
}
\arguments{
\item{XY}{Matrix/data frame with projected coordinates.}

\item{geometry}{Gemetry for spatial weight matrix: "circle", "ellipse", "square", 
"diamond"}

\item{d}{figure distance.}

\item{a}{x radius (semimajor axis)}

\item{b}{y radius (seminimor axis)}
}
\description{
Spatial weights for individuals with XY coordinates
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{internal}
