% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.image2df}
\alias{aue.image2df}
\title{Transforming a raster into a data frame with cartesian coordinates}
\usage{
aue.image2df(mat)
}
\arguments{
\item{mat}{Input raster matrix.}
}
\description{
This function returns a data frame with the column number (x),
row number (y) and cell value (z) of each pixel in a raster.
}
\examples{
\dontrun{

ras <- matrix(eco[["P"]][,1],15,15)
image(ras)
ras.row <- aue.image2df(ras)
ras.row
image(matrix(ras.row[,3], 15, 15))

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

