% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.is.poly}
\alias{aue.is.poly}
\title{Identification of polymorphic loci}
\usage{
aue.is.poly(X, poly.level)
}
\arguments{
\item{X}{allelic frequencies matrix}

\item{poly.level}{polymorphism threshold}
}
\description{
Identification of polymorphic loci
}
\keyword{internal}

