% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.phenosimil}
\alias{aue.phenosimil}
\title{Phenotypic similarity for vector, matrix or data frame acoording to Ritland (1996)}
\usage{
aue.phenosimil(X)
}
\arguments{
\item{X}{Data frame, matrix or vector. If X is not a vector, the program
returns a list of matrices.}
}
\description{
Phenotypic similarity for vector, matrix or data frame acoording to Ritland (1996)
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
