% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.point}
\alias{aue.point}
\title{Solid circle}
\usage{
aue.point(mat, r, x0, y0)
}
\arguments{
\item{mat}{Input raster matrix.}

\item{r}{Radius of the circle in pixels.}

\item{x0}{Circle center- x position.}

\item{y0}{Circle center- y position.}
}
\description{
Solid circle
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}
