% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.rm.nonpoly}
\alias{aue.rm.nonpoly}
\title{Remotion of non polymorphic loci}
\usage{
aue.rm.nonpoly(X, poly.level)
}
\arguments{
\item{X}{allelic frequencies matrix}

\item{poly.level}{polymorphism threshold}
}
\description{
Remotion of non polymorphic loci
}
\keyword{internal}

