% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.rmspaces}
\alias{aue.rmspaces}
\title{Remove spaces and tabs at the begining and the end of each element of charvec}
\usage{
aue.rmspaces(charvec)
}
\arguments{
\item{charvec}{character vector}
}
\description{
Remove spaces and tabs at the begining and the end of each element of charvec
}
\keyword{internal}
