% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.rotate}
\alias{aue.rotate}
\title{Rotation of a matrix}
\usage{
aue.rotate(x, direction = c(1, 2, 3, 4))
}
\arguments{
\item{x}{input matrix}

\item{direction}{vector with four coordinates indicating the position of the matrix angles. 
1 = upper left, 2= upper right, 3 = lower right, 4 = lower left.}

\item{y0}{point column}
}
\description{
The program flips the matrix in the desired direction, given as a vector with four 
coordinates indicating the matrix corners (1 = upper left, 2= upper right, 3 = lower right, 4 = lower left).
The position of the input matrix is c(1,2,3,4). For example, a transposition is c(1,4,3,2).
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

