% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.seqlist}
\alias{aue.seqlist}
\title{Creation of a sequence of numbers in matrix or list format, for indexing.}
\usage{
aue.seqlist(from, to, by, out.format = c("matrix", "list"))
}
\arguments{
\item{from}{start of sequence}

\item{to}{end of sequence}

\item{by}{interval between elements}

\item{out.output}{format ("matrix2 or "list")}
}
\description{
Creation of a sequence of numbers in matrix or list format, for indexing.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

