% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.split_categorical}
\alias{aue.split_categorical}
\title{Split categorical variable into levels,  using a second factor (hierarchy) to aggregate the data}
\usage{
aue.split_categorical(X, hier)
}
\arguments{
\item{X}{factor}

\item{hier}{hierarchy}
}
\description{
Split categorical variable into levels,  using a second factor (hierarchy) to aggregate the data
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{export}
