% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.square}
\alias{aue.square}
\title{Solid square}
\usage{
aue.square(mat, d, x0, y0, theta = 0)
}
\arguments{
\item{mat}{Input raster matrix.}

\item{d}{Radius of the square in pixels.}

\item{x0}{Square center in the x direction.}

\item{y0}{Square center in the y direction.}

\item{theta}{rotation angle in degrees.}
}
\description{
Solid square
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

