% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{E}
\alias{E}
\title{E}
\usage{
data(eco4)
E
}
\description{
data frame with simulated environmental data of 173 individuals.
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{data}
