% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eco.2columns.R
\name{eco.2columns}
\alias{eco.2columns}
\title{Converting a diploid ecogen genetic data frame with two alleles per column
into a data frame with one allele per column}
\usage{
eco.2columns(eco, ndig)
}
\arguments{
\item{eco}{Object of class "ecogen".}

\item{ndig}{Number of digits coding each allele
(e.g., 1: x, 2: xx, or 3: xxx).}
}
\description{
This function converts a diploid ecogen genetic data frame
with two alleles per column into a data frame with one allele per column.
}
\examples{
\dontrun{

data(eco3)
eco.2al <- eco.2columns(eco3, 1)
eco.2al
}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

