% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eco.2gstudio.R
\name{eco.2gstudio}
\alias{eco.2gstudio}
\title{Converting an ecogen genetic data frame into a gstudio object.}
\usage{
eco.2gstudio(eco, type = "separated", ...)
}
\arguments{
\item{eco}{ecogen object}

\item{type}{The type of data passed to \code{\link[gstudio]{locus}}.
Default is "separated" (data as microsatellites or individual haplotypes);
"aflp" for presence - absence data.}

\item{...}{Further arguments passed to \code{\link[adegenet]{df2genind}}}
}
\description{
Converting an ecogen genetic data frame into a gstudio object.
}
\examples{
\dontrun{

data(eco.test)
gsteco <- eco.2gstudio(eco, "separated")
gsteco

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

