% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eco.append.R
\name{eco.append}
\alias{eco.append}
\title{Creating an updated ecogen object by adding
results to the slot @OUT}
\usage{
eco.append(eco, ...)
}
\arguments{
\item{eco}{Object of class "ecogen".}

\item{...}{Objects to include in eco, typed without quotations.}
}
\description{
Creating an updated ecogen object by adding
results to the slot @OUT
}
\examples{
\dontrun{

data(eco.test)
variog <- eco.variogram(eco$P[, 1], eco$XY, int=1, smax = 8)
we.are.integers <- c(1:10)
we.are.characters <- c("John Coltrane", "Charlie Parker")
eco <- eco.append(eco, variog, we.are.integers, we.are.characters)
eco
}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

