% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show_summary.methods.R
\docType{methods}
\name{show,eco.lsa-method}
\alias{show,eco.autol-method}
\alias{show,eco.lsa-method}
\title{show eco.lsa}
\usage{
\S4method{show}{eco.lsa}(object)
}
\description{
show eco.lsa
}
\keyword{internal}

