% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.clear.R
\name{eco.clear}
\alias{eco.clear}
\title{Clearing the working environment, maintaining only the specified objects}
\usage{
eco.clear(..., all = FALSE)
}
\arguments{
\item{...}{Objects to retain.}

\item{all}{Remove also hidden elements? Default FALSE.}
}
\description{
This function removes all the elements of the working environment,
with the exception of those included in the argument of the function. 
Hidden elements can also be removed by setting all = TRUE.
}
\examples{

\dontrun{

data(eco.test)
ls()
eco.clear(eco)
ls()

}

}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
