% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{eco.correlog-class}
\alias{eco.correlog-class}
\title{eco.correlog-class}
\description{
eco.correlog-class
}
\section{Slots}{

\describe{
\item{\code{OUT}}{analysis output}

\item{\code{IN}}{analysis input data}

\item{\code{BEAKS}}{breaks}

\item{\code{CARDINAL}}{number of elements in each class}

\item{\code{NAMES}}{variables names}

\item{\code{METHOD}}{analysis method}

\item{\code{DISTMETHOD}}{method used in the construction of breaks}

\item{\code{TEST}}{test method used (bootstrap, permutation)}

\item{\code{NSIM}}{number of simulations}

\item{\code{PADJUST}}{P-values adjust method for permutation tests}

\item{\code{ANGLE}}{angle used for the construction of correlogram}

\item{\code{BEARING}}{bearing correlogram (logical)}
}}

\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{internal}
