% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{eco.detrend-class}
\alias{eco.detrend-class}
\title{eco.detrend class}
\description{
eco.detrend class
}
\section{Slots}{

\describe{
\item{\code{POLY.DEG}}{polynomial degree used in the analysis}

\item{\code{RES}}{detrended data}

\item{\code{XY}}{projected coordinates}

\item{\code{MODEL}}{models selected with the Akaike criterion}

\item{\code{ANALYSIS}}{object of class "eco.mlm" with 
the regression results for each variable}
}}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

