% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{eco.gsa-class}
\alias{eco.gsa-class}
\title{eco.gsa class}
\description{
eco.gsa class
}
\section{Slots}{

\describe{
\item{\code{METHOD}}{method used in the analysis}

\item{\code{OBS}}{observed value when a single variable is tested}

\item{\code{EXP}}{expected value when a single variable is tested}

\item{\code{PVAL}}{P-value when a single variable is tested}

\item{\code{ALTER}}{alternative hypotesis when a single variable is tested}

\item{\code{NSIM}}{number of simulations}

\item{\code{MULTI}}{table with observed and expected values, P-values and alternative
hypoteses when multiple variables are tested}
}}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

