% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.methods.R
\docType{methods}
\name{plot,eco.IBD,ANY-method}
\alias{plot,eco.IBD,ANY-method}
\alias{plot,eco.IBD-method}
\title{plot eco.IBD}
\usage{
\S4method{plot}{eco.IBD,ANY}(x)
}
\description{
plot eco.IBD
}
\keyword{internal}
