% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.kin.hardy.R
\name{eco.kin.hardy}
\alias{eco.kin.hardy}
\title{Kinship and relationship estimation for dominant markers}
\usage{
eco.kin.hardy(x, fi)
}
\arguments{
\item{x}{ecogen, genind matrix or data.frame. In case of matrix or data frame,
tha data consists in a table with individuals in rows and allele counts 
in columns.}

\item{fi}{assumed Fi}
}
\value{
a list with three slots containing, respectively, the heretability 
of each locus, relationship and kinship values
}
\description{
Kinship and relationship estimation following Hardy (2003).
}
\author{
Juan Vilardi
}
