% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.kin.loiselle.R
\name{eco.kin.loiselle}
\alias{eco.kin.loiselle}
\title{Obtention of the multilocus Loiselle's Fij matrix}
\usage{
eco.kin.loiselle(eco)
}
\arguments{
\item{eco}{Object of class ecogen.}
}
\description{
Obtention of the multilocus Loiselle's Fij matrix
}
\examples{

\dontrun{

data(eco.test)
loiselle <- eco.kin.loiselle(eco)
loiselle[1:5, 1:5]

}

}
\references{
Kalisz, S., J. Nason, F.M. Handazawa, and S. Tonsor. 2001. 
Spatial population genetic structure in Trillium grandiflorum: 
the roles of dispersal, mating, history, and selection. 
Evolution 55: 1560-1568.

Loiselle, B., V. Sork, J. Nason, and C. Graham. 1995. 
Spatial genetic structure of a tropical understory shrub, 
Psychotria officinalis (Rubiaceae). 
American Journal of Botany 1420-1425.
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
