% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eco.laglistw.R
\name{eco.laglistw}
\alias{eco.laglistw}
\title{Obtention of a list with weight matrices for lag distance classes}
\usage{
eco.laglistw(xy, int, smax, w = c("W", "B"))
}
\arguments{
\item{xy}{data frame or matrix with coordinates.}

\item{int}{Lag class interval distance in meters.}

\item{smax}{Maximum lag distance in meters.}

\item{w}{Weight style. "W" for row standardized (default) or "B"
for binary.}
}
\description{
This program returns a list with weights matrices
for each lag distance class.
}
\examples{
\dontrun{
data(eco.test)
lagged <- eco.laglistw(eco$XY, 100, 500)
}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

