% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{eco.lagweight-class}
\alias{eco.lagweight-class}
\title{eco.lagweight class}
\description{
eco.lagweight class
}
\section{Slots}{

\describe{
\item{\code{PAR}}{parameters used for the construction of breaks}

\item{\code{PAR.VAL}}{values of the parameters used for the construction of breaks}

\item{\code{ROW.SD}}{row standardization (logical)}

\item{\code{SELF}}{data self-included (logical)}

\item{\code{W}}{weights list}

\item{\code{XY}}{input coordinates}

\item{\code{MEAN}}{mean class distances}

\item{\code{LOGMEAN}}{mean of the class distances logarithm}

\item{\code{CARDINAL}}{number of elements in each class}

\item{\code{BREAKS}}{breaks}

\item{\code{METHOD}}{breaks construction method}

\item{\code{ANGLE}}{angle used for the construction of correlogram}
}}

\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}
