% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show_summary.methods.R
\docType{methods}
\name{show,eco.lagweight-method}
\alias{show,eco.lagweight-method}
\title{show eco.lagweight}
\usage{
\S4method{show}{eco.lagweight}(object)
}
\description{
show eco.lagweight
}
\keyword{internal}

