% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.methods.R
\docType{methods}
\name{plot,eco.listlsa,ANY-method}
\alias{plot,eco.listlsa,ANY-method}
\alias{plot,eco.listlsa-method}
\title{listplot}
\usage{
\S4method{plot}{eco.listlsa,ANY}(x, significant = TRUE, rescaled = FALSE,
  legend = TRUE, n = 4, nrow = 2, byrow = TRUE, ...)
}
\arguments{
\item{x}{list of ggplot objects}

\item{significant}{plot only significant individuals?}

\item{n}{number of plot in layout}

\item{nrow}{Number of rows in layout}

\item{byrow}{plot by row?}
}
\description{
Plot method for local spatial analysis as list
}
\examples{
\dontrun{

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\seealso{
\code{\link{eco.lsa}}
}
\keyword{internal}

