% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{eco.listw2ew}
\alias{eco.listw2ew}
\title{Conversion from listw to ecoweight}
\usage{
eco.listw2ew(X)
}
\arguments{
\item{X}{A listw object}
}
\description{
Conversion from listw to ecoweight
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
