% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/show_summary.methods.R
\docType{methods}
\name{summary,eco.mlm-method}
\alias{summary,eco.lmtree-method}
\alias{summary,eco.mctree-method}
\alias{summary,eco.mlm-method}
\title{Summary for eco.lmtree output}
\usage{
\S4method{summary}{eco.mlm}(object)

\S4method{summary}{eco.mctree}(object)
}
\arguments{
\item{object}{Output object of \code{\link{eco.lmtree}}.}
}
\value{
A Table with a summary of the analysis for "mlm" analysis,
the plot of the trees with significant splits for "mctree" analysis.
}
\description{
Summary for eco.lmtree output
}
\examples{
\dontrun{

data(eco.test)
#' mod <- eco.lmtree(df1 = eco$P, df2 = eco$E,
analysis = "mlm")
summary(mod)                                    #summary for "mlm" analysis

mod <- eco.lmtree(df1 = eco$P, df2 = eco$E,
analysis = "mctree", fact = eco$S$structure)
summary(mod)                                    #summary for "mctree" analysis

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\seealso{
\code{\link{eco.lmtree}}
}

