% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopop.3OF6.basic.methods.R
\name{eco.lock_rows}
\alias{eco.lock_rows}
\title{Lock rows in an ecogen object}
\usage{
eco.lock_rows(object, set.names, valid.names)
}
\arguments{
\item{set.names}{Character vector with names for the rows of the non-empty data frames. 
This argument is incompatible with valid.names}

\item{valid.names}{Logical. Create valid row names? This argument is incompatible with 
set.names. The program will name individuals with valid tags I.1, I.2, etc.}
}
\description{
This methods locks the rows in an ecogen object.  When rows are locked,
the object requires rows with identical indviduals in the non empty data frames, and
identity in the row names of the data frames.
}
