% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{eco.lsa-class}
\alias{eco.lsa-class}
\title{eco.lsa class}
\description{
eco.lsa class
}
\section{Slots}{

\describe{
\item{\code{OUT}}{results}

\item{\code{METHOD}}{method used in the analysis}

\item{\code{TEST}}{test method used (bootstrap, permutation)}

\item{\code{NSIM}}{number of simulations}

\item{\code{PADJUST}}{P-values adjust method for permutation tests}

\item{\code{COND}}{conditional randomization (logical)}

\item{\code{XY}}{input coordinates}
}}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

