% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.methods.R
\docType{methods}
\name{plot,eco.lsa,ANY-method}
\alias{plot,eco.lsa,ANY-method}
\alias{plot,eco.lsa-method}
\title{plot eco.lsa}
\usage{
\S4method{plot}{eco.lsa,ANY}(x, rescaled = FALSE, ...)
}
\description{
plot eco.lsa
}
\keyword{internal}

