% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{eco.mctree-class}
\alias{eco.mctree-class}
\title{eco.mctree-class}
\description{
eco.mctree-class
}
\section{Slots}{

\describe{
\item{\code{TREES}}{trees obtained}

\item{\code{PREDICTIONS}}{predictions of the analysis}

\item{\code{FREQUENCIES}}{frequencies of individuals per class in nodes}

\item{\code{df1}}{data frame}

\item{\code{df2}}{data frame}
}}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

