% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{eco.mlm-class}
\alias{eco.mlm-class}
\title{eco.mlm-class}
\description{
eco.mlm-class
}
\section{Slots}{

\describe{
\item{\code{MLM}}{mlm results}

\item{\code{SUMMARY.MLM}}{summary of the results}

\item{\code{ANOVA.MLM}}{anovas for the results}

\item{\code{DF1}}{data frame}

\item{\code{DF2}}{data frame}
}}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

