% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eco.rasterplot.R, R/plot.methods.R
\docType{methods}
\name{eco.rasterplot,eco.multilsa,missing,missing-method}
\alias{eco.rasterplot,eco.multilsa,missing,missing-method}
\alias{plot,eco.multilsa,ANY-method}
\alias{plot,eco.multilsa-method}
\title{rasterplot graph for eco.lsa results}
\usage{
\S4method{eco.rasterplot}{eco.multilsa,missing,missing}(x, grp = NULL,
  limits = NULL, title = NULL, z.name = NULL, vertical = TRUE,
  significant = TRUE, rescaled = FALSE, alpha = 0.05)

\S4method{plot}{eco.multilsa,ANY}(x, significant = TRUE, alpha = 0.05,
  rescaled = FALSE, limits = NULL, title = NULL, z.name = NULL,
  grp = NULL, vertical = TRUE)
}
\arguments{
\item{x}{eco.multilsa object returned by \code{\link{eco.lsa}} or}

\item{limits}{values limits used for computing the data gradient for the plot}

\item{title}{plot title}

\item{z.name}{name for the legend}

\item{significant}{plot only significant results?  Default TRUE}

\item{rescaled}{plot the rescaled observed values ([-1,1] range)?}

\item{alpha}{threshold P value for results with permutation tests. default = 0.05.}
}
\description{
Plot method for local spatial analysis

Plot method for local spatial analysis
}
\examples{
\dontrun{

}
\dontrun{

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}

Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\seealso{
\code{\link{eco.lsa}}

\code{\link{eco.lsa}}
}

