% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eco.n.R
\docType{data}
\name{eco.n}
\alias{eco.n}
\title{Eco.n}
\usage{
data(eco.test)
eco.n
}
\description{
ecogen object with simulated data of 225 individuals.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{data}

