% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.nei_distance.R
\name{eco.nei_distance}
\alias{eco.nei_distance}
\title{Estimate Nei distance matrix}
\usage{
eco.nei_distance(obj, as.dist = TRUE)
}
\arguments{
\item{x}{Matrix of allele frequencies}

\item{as_dist}{Return a dist object or a matrix? default an object of class "dist".}
}
\description{
Estimate Nei distance matrix. NAs are avoided.
}
\author{
Juan Vilardi
}
