% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecogen.3OF6.basic.methods.R
\docType{methods}
\name{eco.old2new,ecogen-method}
\alias{eco.old2new,ecogen-method}
\alias{eco.old2new,ecogen}
\title{Update an old ecogen or ecopop object to version  >= 1.5.0-1}
\usage{
\S4method{eco.old2new}{ecogen}(object)
}
\description{
Update an old ecogen or ecopop object to version  >= 1.5.0-1
}
