% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopop.3OF6.basic.methods.R
\docType{methods}
\name{eco.old2new,ecopop-method}
\alias{eco.old2new,ecopop-method}
\alias{eco.old2new,ecopop}
\title{Update an old ecogen or ecopop object to a version compatible with EcoGenetics >= 1.5.0-1}
\usage{
\S4method{eco.old2new}{ecopop}(object)
}
\arguments{
\item{object}{ecopop object}
}
\description{
Update an old ecogen or ecopop object to a version compatible with EcoGenetics >= 1.5.0-1
}
