% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{eco.order}
\alias{eco.order}
\title{Functions deprecated in EcoGenetics version 1.2.0-2}
\usage{
eco.order(...)
}
\arguments{
\item{...}{parameters}
}
\description{
Functions deprecated in EcoGenetics version 1.2.0-2
}
