% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.plotGlobal.R
\name{eco.plotGlobal}
\alias{eco.plotGlobal}
\title{GSA plot methods}
\usage{
eco.plotGlobal(input, interactivePlot = TRUE, background = c("grey",
  "white"), xlabel = NULL, ylabel = NULL, title = NULL, legend = TRUE,
  rescaled = FALSE, alpha = 0.05)
}
\arguments{
\item{input}{eco.gsa object}

\item{interactivePlot}{Show an interactive plot via plotly? (default: TRUE)}

\item{background}{background color ("grey" or "white")}

\item{xlabel}{Label for X axis (default: NULL)}

\item{ylabel}{Label for Y axis (default: NULL)}

\item{title}{Title of the plot (default: NULL)}

\item{legend}{Show legends in ggplot graphs? (default: TRUE)}

\item{rescaled}{rescale join-count heatmap?}

\item{alpha}{significance level for the join-count heatmat}
}
\description{
This function allows to plot eco.gsa multiple objects.
For examples see \code{\link{eco.gsa}}
}
\author{
Leandro Roser
}

