% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.pop_alfreq.R
\name{eco.pop_afreq}
\alias{eco.pop_afreq}
\title{Compute allele frequencies using different methods}
\usage{
eco.pop_afreq(x, method = c("zhivor", "zhivonu", "rawfreq", "neifreq"))
}
\arguments{
\item{x}{ecopop or genpop object}

\item{method}{Method used to compute the allelic frequencies. 
Can be one of 'zhivor' ( Zhivotovsky (1999) with uniform prior),  
'zhivonu' (Zhivotovsky (1999) with non uniform prior), 
'rawfreq' (square root method),  'neifreq' (nei distance matrix)}
}
\description{
Compute allele frequencies using different methods
}
\author{
Juan Vilardi
}
