% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.rasterplot.R
\name{eco.rasterplot,eco.multilsa-method}
\alias{eco.rasterplot,eco.multilsa-method}
\title{rasterplot graph for eco.lsa results}
\usage{
\S4method{eco.rasterplot}{eco.multilsa}(
  x,
  grp = NULL,
  limits = NULL,
  title = NULL,
  z.name = NULL,
  vertical = TRUE,
  significant = TRUE,
  rescaled = FALSE,
  alpha = 0.05,
  interactivePlot = TRUE,
  ...
)
}
\arguments{
\item{x}{eco.multilsa object returned by \code{\link{eco.lsa}} or}

\item{grp}{factor with groups to use in the plot. Default NULL}

\item{limits}{values limits used for computing the data gradient for the plot}

\item{title}{plot title}

\item{z.name}{name for the legend}

\item{vertical}{should be partitioned the populations on the x axis? Default
TRUE.}

\item{significant}{plot only significant results?  Default TRUE}

\item{rescaled}{plot the rescaled observed values ([-1,1] range)?}

\item{alpha}{threshold P value for results with permutation tests. default = 0.05.}

\item{interactivePlot}{Show an interactive plot via plotly? (default: TRUE)}

\item{...}{additional arguments}
}
\description{
Plot method for local spatial analysis
}
\seealso{
\code{\link{eco.lsa}}
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
