% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eco.rbind.R
\name{eco.rbind}
\alias{eco.rbind}
\title{Combining the rows of two ecogen objects}
\usage{
eco.rbind(e1, e2)
}
\arguments{
\item{e1}{Object of class "ecogen".}

\item{e2}{Object of class "ecogen".}
}
\description{
Combining the rows of two ecogen objects
}
\examples{
\dontrun{

data(eco.test)

# crating a copy of "eco"
eco.n <- eco

# duplicated row names are not allowed by eco.rbind.
# creating new row names for eco.n
names.eco <- rownames(eco$XY)
names.eco <- paste(names.eco, ".n", sep = "")
rownames(eco.n$XY) <- names.eco
rownames(eco.n$P) <- names.eco
rownames(eco.n$G) <- names.eco
rownames(eco.n$E) <- names.eco
rownames(eco.n$S) <- names.eco

# binding by row eco and eco.n
eco2<-eco.rbind(eco, eco.n)

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

