% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eco.remove.R
\name{eco.remove}
\alias{eco.remove}
\title{Creating an updated ecogen object by removing
results of the slot OUT.}
\usage{
eco.remove(object, ...)
}
\arguments{
\item{object}{Ecogen object.}

\item{...}{Objects to remove, typed without quotations.}
}
\description{
Creating an updated ecogen object by removing
results of the slot OUT.
}
\examples{
\dontrun{

data(eco.test)
variog <- eco.variogram(eco, eco@P$P1)
eco <- eco.append(eco, variog)
we.are.numbers <- c(1:10)
we.are.characters <- c("John Coltrane", "Charlie Parker")
eco <- eco.append(eco, we.are.numbers, we.are.characters)
eco
eco <- eco.remove(eco, we.are.numbers)
eco

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

