% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eco.slide.matrix.R
\name{eco.slide.matrix}
\alias{eco.slide.matrix}
\title{Sliding window for matrix data}
\usage{
eco.slide.matrix(mat, r, slide, fun, window = c("square", "circle",
  "rhombus"), within = TRUE)
}
\arguments{
\item{mat}{Input raster matrix.}

\item{r}{half a side for square, radius for circle, diagonal length for rhombus.}

\item{slide}{number of elements between two focal pixels, for column
and row dimension}

\item{fun}{Function to apply in each focal pixel.}

\item{window}{window type. Default "square".}

\item{within}{should be computed the function in borders focal pixels only if
the area is within the matrix? Default TRUE.}
}
\description{
This program applies a function defined by the user,
using a moving window (circle area or square) and assigning
the value to the focal pixel.
}
\examples{
\dontrun{

ras <- matrix(eco[["P"]][,1],15,15)
image(ras)
ras.square <- eco.slidewindow(mat, 1, 1, mean, "diamond")

# or allowing more control over the function:
ras.square <- eco.slidewindow(mat, r = 3, slide = 1, function(x)mean(x, na.rm = TRUE), "diamond")
image(ras.square)

ras.circle <- eco.slidewindow(mat, r = 3, slide = 1, mean, "circle", within = FALSE)
image(ras.circle)

ras.diamond <- eco.slidewindow(mat, r = 3, slide = 1, mean, "diamond")
image(ras.square)
}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

