% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecogen.3OF6.basic.methods.R
\docType{methods}
\name{eco.unlock,ecogen-method}
\alias{eco.unlock,ecogen-method}
\alias{eco.unlock,ecogen}
\title{Unlock rows in an ecogen object}
\usage{
\S4method{eco.unlock}{ecogen}(object)
}
\arguments{
\item{object}{object of class ecogen}
}
\description{
This methods unlocks the rows in an ecogen object. This means that 
different data frames in the object can have different rows, with different row names.
}
\examples{
\dontrun{
data(eco.test)
eco2 <- eco.unlock(eco)
is.locked(eco2) 
eco3 <- eco.lock(eco2)
is.locked(eco3) 
}

}
