% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopop.3OF6.basic.methods.R
\docType{methods}
\name{eco.unlock,ecopop-method}
\alias{eco.unlock,ecopop-method}
\title{Unlock rows in an ecogen object}
\usage{
\S4method{eco.unlock}{ecopop}(object)
}
\description{
This methods unlocks the rows in an ecogen object. This means that 
different data frames in the object can have different rows, with different row names.
}
