% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopop.3OF6.basic.methods.R
\name{eco.unlock_rows}
\alias{eco.unlock_rows}
\title{Unlock rows in an ecogen object}
\usage{
eco.unlock_rows(object)
}
\description{
This methods unlocks the rows in an ecogen object. This means that 
different data frames in the object can have different rows, with different row names.
}
