% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{eco.variogram-class}
\alias{eco.variogram-class}
\title{eco.variogram class}
\description{
eco.variogram class
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

