% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{eco.weight-class}
\alias{eco.weight-class}
\title{eco.weight class}
\description{
eco.weight class
}
\section{Slots}{

\describe{
\item{\code{METHOD}}{weights construction method}

\item{\code{PAR}}{parameters used for the construction of breaks}

\item{\code{PAR.VAL}}{values of the parameters used for the construction of breaks}

\item{\code{ROW.SD}}{row standardization (logical)}

\item{\code{SELF}}{data self-included (logical)}

\item{\code{W}}{weights list}

\item{\code{XY}}{input coordinates}

\item{\code{NONZERO}}{number non-zero connections}

\item{\code{NONZEROIND}}{number of individuals}

\item{\code{NONZEROINDROW}}{individuals with non-null connections
with non-zero connections (as percentage)}

\item{\code{AVERAGE}}{average number of connection (as percentage)}

\item{\code{AVG.DIST}}{average distance between connected individuals}

\item{\code{ANGLE}}{angle used for the construction of correlogram}
}}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

