% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_summary.methods.R
\docType{methods}
\name{show,eco.weight-method}
\alias{show,eco.weight-method}
\title{show eco.weight}
\usage{
\S4method{show}{eco.weight}(object)
}
\description{
show eco.weight
}
\keyword{internal}

