% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eco2.R
\docType{data}
\name{eco2}
\alias{eco2}
\title{Eco2}
\usage{
data(eco2)
eco2
}
\description{
ecogen object with simulated data of 900 individuals.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{data}

