% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{eco3}
\alias{eco3}
\title{eco3}
\usage{
data(eco3)
eco3
}
\description{
ecogen object with simulated data of 173 individuals.
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{data}
