% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{eco4}
\alias{eco4}
\title{Eco4}
\usage{
data(eco3)
eco3
}
\description{
data frames with simulated data of 173 individuals. 
The data frams can be used to construct an ecogen object that includes
genetic data separated by a character ad with non uniform of number of characters
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{data}
