% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecogen.1OF6.definition.R
\docType{class}
\name{ecogen-class}
\alias{ecogen-class}
\title{ecogen class}
\description{
ecogen class
}
\section{Slots}{

\describe{
\item{\code{XY}}{P data frame}

\item{\code{P}}{P data frame}

\item{\code{G}}{G data frame}

\item{\code{A}}{A matrix}

\item{\code{E}}{E data frame}

\item{\code{S}}{S data frame}

\item{\code{C}}{C data frame}

\item{\code{OUT}}{results}

\item{\code{INT}}{int.gendata slot}

\item{\code{ATTR}}{attributes slot}
}}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{internal}

