% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecogen.3OF6.basic.methods.R, R/ecogen.4OF6.brackets.R
\docType{methods}
\name{names,ecogen-method}
\alias{$,ecogen,character-method}
\alias{$,ecogen-method}
\alias{$<-,ecogen,character,ANY-method}
\alias{$<-,ecogen-method}
\alias{[,ecogen,character,missing-method}
\alias{[,ecogen,logical,missing,ANY-method}
\alias{[,ecogen,logicalORmissing,missing,ANY-method}
\alias{[,ecogen,numeric,missing,ANY-method}
\alias{[,ecogen,numericORmissing,missing,ANY-method}
\alias{[[,ecogen,character,missing-method}
\alias{[[<-,ecogen,character,missing-method}
\alias{as.int.list}
\alias{as.list,ecogen-method}
\alias{dim,ecogen-method}
\alias{is,ecogen-method}
\alias{is.ecogen}
\alias{names,ecogen-method}
\alias{names<-,ecogen,character-method}
\alias{ncol,ecogen-method}
\alias{nrow,ecogen-method}
\alias{show,ecogen-method}
\title{names}
\usage{
\S4method{names}{ecogen}(x)

\S4method{names}{ecogen,character}(x) <- value

is.ecogen(x)

\S4method{nrow}{ecogen}(x)

\S4method{ncol}{ecogen}(x)

\S4method{dim}{ecogen}(x)

\S4method{as.list}{ecogen}(x)

\S4method{as.int.list}{ecogen}(x)

\S4method{show}{ecogen}(object)

\S4method{$}{ecogen}(x, name)

\S4method{$}{ecogen}(x, name) <- value

\S4method{[}{ecogen,numericORmissing,missing,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{ecogen,logicalORmissing,missing,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{ecogen,character,missing}(x, i, j)

\S4method{[[}{ecogen,character,missing}(x, i, j, ...) <- value
}
\arguments{
\item{x}{Object of class ecogen}

\item{x}{Object of class ecogen}
}
\description{
names

names<-

is.ecogen

nrow

ncol

dim

as.list

as.int.list

show method

$

$<-

[

[

[[

[[<- -----------------------------------------------------------------------#
}
\keyword{internal}

