% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ecogen.definition.R
\docType{methods}
\name{summary,ecogen-method}
\alias{summary,ecogen,ANY-method}
\alias{summary,ecogen-method}
\title{Summary for ecogen objects}
\usage{
\S4method{summary}{ecogen}(object, x = NULL)
}
\arguments{
\item{object}{Ecogen object.}

\item{x}{The name of the S slot column with the grouping factor
when a summary taking in account groups is required.}
}
\value{
Redundancy analysis for the P data frame in function of G
(constraining data frame)  and E (unconstraining data frame).

Mean and sd table for the P data frame.

Genetic analysis for the G data frame via hierfstat. This analysis
can be performed when the G data frame has an appropriate hierfstat
input format (i.e., columns of class numeric or integer, neither factor nor character).
}
\description{
Summary for ecogen objects
}
\examples{
\dontrun{

data(eco.test)
summary(eco, "structure")

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\seealso{
\code{\link{rda}} \code{\link{basic.stats}}
}

