% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopop.6OF6.converters.R
\name{ecogen2ecopop}
\alias{ecogen2ecopop}
\title{Conversion form ecogen to ecopop}
\usage{
ecogen2ecopop(from, hier, factor_to_counts = TRUE, aggregator = function(x)
  mean(x, na.rm = TRUE))
}
\arguments{
\item{from}{Object of class "ecogen"}

\item{hier}{Name of the level of the slot S with hierarchies}

\item{factor_to_counts}{Convert factors into counts for each level?}

\item{aggregator}{Function used to aggregate data}
}
\description{
This function creates an ecopop object from an ecogen object
}
\examples{

\dontrun{

data(eco.test)
ecogen2ecopop(eco, hier = "pop")

}

}
\author{
Leandro Roser \email{learoser@gmail.com}
}
