% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecogen.6OF6.converters.R
\name{ecogen2genind}
\alias{ecogen2genind}
\alias{genind2ecogen}
\title{Conversion form ecogen to genind and genind to ecogen}
\usage{
ecogen2genind(from)

genind2ecogen(from)
}
\arguments{
\item{from}{Object of class "ecogen" / "genind"}
}
\description{
These functions export from ecogen to genind and viceversa

genind2ecogen
}
\examples{

\dontrun{

data(eco.test)

# ecogen to genind
outGenind <- ecogen2genind(eco)
outGenind

# genind to ecogen
outEco <- genind2ecogen(outGenind)

}

}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
