% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{ecogenetics_devel}
\alias{ecogenetics_devel}
\title{EcoGenetic devel site}
\usage{
ecogenetics_devel()
}
\description{
The function opens the EcoGenetics-devel web site:
https://github.com/leandroroser/EcoGenetics-devel
}
