% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{ecogenetics_tutorial}
\alias{ecogenetics_tutorial}
\title{EcoGenetic tutorial site}
\usage{
ecogenetics_tutorial()
}
\description{
The function opens the EcoGenetics tutorial web site: 
https://leandroroser.github.io/EcoGenetics-Tutorial
}
