% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopop.1OF6.definition.R
\docType{class}
\name{ecopop-class}
\alias{ecopop-class}
\title{ecopop class}
\description{
ecopop class
}
\section{Slots}{

\describe{
\item{\code{XY}}{P data frame}

\item{\code{P}}{P data frame}

\item{\code{AF}}{AF data frame}

\item{\code{E}}{E data frame}

\item{\code{S}}{S data frame}

\item{\code{C}}{C data frame}

\item{\code{INT}}{int.popdata slot}

\item{\code{ATTR}}{attributes slot}
}}

\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{internal}
