% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopop.3OF6.basic.methods.R,
%   R/ecopop.4OF6.brackets.R
\docType{methods}
\name{names,ecopop-method}
\alias{names,ecopop-method}
\alias{names<-,ecopop,any_vector-method}
\alias{names,ecopop-method}
\alias{is.ecopop}
\alias{is,ecopop-method}
\alias{nrow,ecopop-method}
\alias{ncol,ecopop-method}
\alias{dim,ecopop-method}
\alias{as.list,ecopop-method}
\alias{show,ecopop-method}
\alias{$,ecopop-method}
\alias{$,ecopop,character-method}
\alias{$<-,ecopop-method}
\alias{$<-,ecopop,character,ANY-method}
\alias{[,ecopop,numericORmissing,missing,ANY-method}
\alias{[,ecopop,numeric,missing,ANY-method}
\alias{[,ecopop,logicalORmissing,missing,ANY-method}
\alias{[,ecopop,logical,missing,ANY-method}
\alias{[[,ecopop,character,missing-method}
\alias{[[<-,ecopop,character,missing-method}
\alias{[,ecopop,character,missing-method}
\title{names}
\usage{
\S4method{names}{ecopop}(x)

\S4method{names}{ecopop,any_vector}(x) <- value

is.ecopop(x)

\S4method{nrow}{ecopop}(x)

\S4method{ncol}{ecopop}(x)

\S4method{dim}{ecopop}(x)

\S4method{as.list}{ecopop}(x)

\S4method{show}{ecopop}(object)

\S4method{$}{ecopop}(x, name)

\S4method{$}{ecopop}(x, name) <- value

\S4method{[}{ecopop,numericORmissing,missing,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{ecopop,logicalORmissing,missing,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{ecopop,character,missing}(x, i, j)

\S4method{[[}{ecopop,character,missing}(x, i, j, ...) <- value
}
\arguments{
\item{x}{Object of class ecopop}
}
\description{
names

is.ecopop

nrow

ncol

dim

as.list

show method

$

$<-

[

[

[[

[[<-
}
\author{
Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}

Leandro Roser \email{learoser@gmail.com}
}
\keyword{-----------------------------------------------------------------------#}
\keyword{[[<-}
\keyword{internal}
