% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopop.6OF6.converters.R
\name{ecopop_counts2af}
\alias{ecopop_counts2af}
\title{ecopop_counts2af}
\usage{
ecopop_counts2af(from)
}
\description{
Conversion from ecopop with genetic data as count,
into ecopop with genetic data as allele frequencies
}
