% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\docType{data}
\name{environment}
\alias{environment}
\title{environment}
\usage{
data(eco.test)
environment
}
\description{
Data frame with simulated environmental variables
of 225 individuals.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{data}

