% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{G}
\alias{G}
\title{G}
\usage{
data(eco4)
G
}
\description{
data frame with simulated genetic data of 173 individuals.
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{data}
