% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{genotype_dom}
\alias{genotype_dom}
\title{genotype_dom}
\usage{
data(eco.test)
genotype_dom
}
\description{
Data frame with simulated dominant data
data of 225 individuals.
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{data}
