% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.generic.R
\name{grf.seqmultiplot}
\alias{grf.seqmultiplot}
\title{Plot a ggplot sequence in layers of n plots arranged in k rows}
\usage{
grf.seqmultiplot(x, n, nrow, byrow = TRUE)
}
\arguments{
\item{x}{list of ggplot objects}

\item{n}{number of plot in layout}

\item{nrow}{Number of rows in layout}

\item{byrow}{plot by row?}
}
\description{
Plot a ggplot sequence in layers of n plots arranged in k rows
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
