% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{int.2symmetric}
\alias{int.2symmetric}
\title{Conversion of a non symmetric binary matrix into symmetric.}
\usage{
int.2symmetric(mat)
}
\arguments{
\item{mat}{Matrix.}
}
\description{
Conversion of a non symmetric binary matrix into symmetric.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

