% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{int.check.colnames}
\alias{int.check.colnames}
\title{Check column names}
\usage{
int.check.colnames(X, lab = "L")
}
\arguments{
\item{X}{Matrix or data frame.}

\item{lab}{String used as label.}
}
\description{
Check column names
}
\keyword{internal}
