% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliar.R, R/control.R
\name{int.check.con}
\alias{int.check.con}
\title{Checks a connection network. If "con" is of class "listw",
it returns a matrix}
\usage{
int.check.con(con)

int.check.con(con)
}
\arguments{
\item{con}{Connection network.}

\item{con}{Connection network.}
}
\description{
Checks a connection network. If "con" is of class "listw",
it returns a matrix

check.con internal
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

