% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/control.R
\name{int.check.numeric}
\alias{int.check.numeric}
\title{int.check.numeric}
\usage{
int.check.numeric(mat)
}
\arguments{
\item{x}{matrix or data frame}
}
\description{
int.check.numeric
}
\keyword{internal}

