% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{int.check.rownames}
\alias{int.check.rownames}
\title{Check row names}
\usage{
int.check.rownames(X, lab = "")
}
\arguments{
\item{X}{Matrix or data frame.}

\item{lab}{String used as label.}
}
\description{
Check row names
}
\keyword{internal}
