% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{int.check.to_numeric}
\alias{int.check.to_numeric}
\title{check if elements are numbers and if not convert the matrix using eco.format}
\usage{
int.check.to_numeric(x, columns_to_numeric = FALSE, ...)
}
\description{
check if elements are numbers and if not convert the matrix using eco.format
}
\keyword{internal}
