% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{int.check.vnames}
\alias{int.check.vnames}
\title{Check a vector of names}
\usage{
int.check.vnames(X, len.X, lab = "V")
}
\arguments{
\item{X}{Vector of names.}

\item{len.X}{Expected length of X.}

\item{lab}{String used as label.}
}
\description{
Check a vector of names
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}
