% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/int.crosscor.R
\name{int.crosscor}
\alias{int.crosscor}
\title{Cross correlation. Internal.}
\usage{
int.crosscor(Z, Y, con, nsim, alternative, test = "permutation",
  adjust.n = FALSE, plotit)
}
\arguments{
\item{Z}{vector, matrix or data frame.
with variable/s (in matrix or data frame formats, variables in columns).}

\item{Y}{Vector with the second variable for cross-correlograms construction.
if Z has multiple variables, the program will compute the cross correlograms
between each and Y.}

\item{con}{Connection network.}

\item{nsim}{Number of Monte-Carlo simulations.}

\item{alternative}{The alternative hypothesis. If "auto" is selected (default) the
program determines the hypothesis by difference between the median of the simulations
and the observed value. Other options are: "two.sided", "greater" and "less".
if test == cross, for the first interval (d== 0) the p and CI are computed with cor.test.}

\item{adjust.n}{Should be adjusted the number of individuals? (warning, this would
change variances)}

\item{plotit}{Should be generated a plot of the simulations?}
}
\description{
Cross correlation. Internal.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

