% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/int.genind.R
\name{int.df2genind}
\alias{int.df2genind}
\title{importer}
\usage{
int.df2genind(indata, sep = "", ncod = NULL, NA.char = "NA", ploidy = 2,
  type = c("codominant", "dominant"), missing = c("0", "NA", "MEAN"),
  rm.empty.ind = FALSE, poly.level = 5)
}
\description{
importer
}
\keyword{internal}

