% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{int.dlatlon2distm}
\alias{int.dlatlon2distm}
\title{Computing a distance matrix in meters among points in decimal degrees
under a spherical Earth model}
\usage{
int.dlatlon2distm(XY)
}
\arguments{
\item{XY}{data frame or matrix with latitude-longitude coordinates
in decimal degrees format.
This program computes a distance matrix for Earth points in decimal degrees.
It assumes a spherical model with an Earth radius of 6371 km.}
}
\description{
Computing a distance matrix in meters among points in decimal degrees
under a spherical Earth model
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

