% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.geary.R
\name{int.geary}
\alias{int.geary}
\title{Geary internal.}
\usage{
int.geary(Z, con, nsim, alternative, test = "permutation", adjust.n = FALSE,
  plotit)
}
\arguments{
\item{Z}{Vector, matrix or data frame.}

\item{con}{Connection network.}

\item{nsim}{Number of Monte-Carlo simulations.}

\item{alternative}{The alternative hypothesis. If "auto" is selected (default) the
program determines the hypothesis by difference between the median of the simulations
and the observed value. Other options are: "two.sided", "greater" and "less".
if test == cross, for the first interval (d== 0) the p and CI are computed with cor.test.}

\item{adjust.n}{Should be adjusted the number of individuals? (warning, this would
change variances)}

\item{plotit}{Should be generated a plot of the simulations?}
}
\description{
Geary internal.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

