% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/int.genind.R
\docType{class}
\name{int.gendata-class}
\alias{int.gendata-class}
\title{int.gendata}
\description{
int.gendata
}
\section{Slots}{

\describe{
\item{\code{loc.fac}}{locus of each allele}

\item{\code{all.names}}{alleles names}

\item{\code{ploidy}}{ploidy}

\item{\code{type}}{type of data ("codominant" or "dominant")}

\item{\code{NA.char}}{NA character}

\item{\code{sep}}{separator}

\item{\code{ncod}}{number of digits coding each allele (codominant data)}

\item{\code{missing}}{missing character}

\item{\code{missing}}{cells character}

\item{\code{removed.image}}{removed columns (coded as 2) or rows (coded as 1)}
}}
\keyword{internal}

