% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/int.genind.R
\name{int.gendata2genind}
\alias{int.gendata2genind}
\title{int.gendata2genind}
\usage{
int.gendata2genind(tab, input)
}
\arguments{
\item{tab}{frequency matrix}

\item{input}{int.gendata object}
}
\description{
int.gendata2genind
}
\keyword{internal}

