% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/int.genind.R
\name{int.genind}
\alias{int.genind}
\title{constructor}
\usage{
int.genind(X, ploidy = 2, type = c("codominant", "dominant"),
  NA.char = "NA", missing = c("0", "NA", "MEAN"),
  missing.cells = integer(0), sep = "", ncod = NULL, removed.image)
}
\description{
constructor
}
\keyword{internal}

