% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.genind.R
\name{int.genind2gendata}
\alias{int.genind2gendata}
\title{int.genind2gendata}
\usage{
int.genind2gendata(input)
}
\arguments{
\item{input}{int.genind object}
}
\description{
int.genind2gendata
}
\keyword{internal}

