% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/int.jackknife.R
\name{int.jackknife}
\alias{int.jackknife}
\title{Jackknife estimation.}
\usage{
int.jackknife(x, fun)
}
\arguments{
\item{x}{vector with simulations or matrix with variables in columns, simulations in rows.}

\item{fun}{function to apply.}
}
\description{
Jackknife estimation.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

