% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.joincount.R
\name{int.joincount}
\alias{int.joincount}
\title{Join-count statistic, internal.}
\usage{
int.joincount(Z, con, nsim, alternative, test = "permutation",
  adjust.n = FALSE, adjustjc = "none", plotit)
}
\arguments{
\item{Z}{Vector, matrix or data frame.}

\item{con}{Connection network.}

\item{nsim}{Number of Monte-Carlo simulations.}

\item{alternative}{The alternative hypothesis. If "auto" is selected (default) the
program determines the hypothesis by difference between the median of the simulations
and the observed value. Other options are: "two.sided", "greater" and "less".
if test == cross, for the first interval (d== 0) the p and CI are computed with cor.test.}

\item{test}{If test = "bootstrap", the program generates a bootstrap 
resampling and the associated confidence intervals of the null hypothesis.
 If test = "permutation" (default) a permutation test is made and the p value 
 is calculated.}

\item{adjust.n}{Should be adjusted the number of individuals? (warning, this would
change variances)}

\item{adjustjc}{Method for multiple correction of P-values 
passed to \code{\link[stats]{p.adjust}}.}
}
\description{
Join-count statistic, internal.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

