% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/int.kin.loiselle.R
\name{int.kin.loiselle}
\alias{int.kin.loiselle}
\title{obtetion of multilocus Loiselle's Fij matrix}
\usage{
int.kin.loiselle(geno, locus)
}
\arguments{
\item{eco}{Object of class ecogen.}
}
\description{
obtetion of multilocus Loiselle's Fij matrix
}
\keyword{internal}

