% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/int.convert.R
\name{int.loc2al}
\alias{int.loc2al}
\title{INTERNAL CONVERSION TOOLS FOR GENETIC DATA}
\usage{
int.loc2al(X, ncod = NULL, ploidy = 2, sep.in = "", sep.out = NULL,
  chk.names = TRUE, chk.plocod = TRUE)
}
\arguments{
\item{X}{Input}

\item{ncod}{Number of digits coding each allele
(e.g., 1: x, 2: xx, 3: xxx, etc.). If NULL, ncod will we
obtained from the ploidy and the maximum number of characters
in the data cells.}

\item{ploidy}{Ploidy of the data.}

\item{sep.in}{Separator in the input}

\item{sep.out}{Separator in the output}

\item{chk.names}{Defalult TRUE. The function make checks of individuals
and loci names during conversion.}

\item{chk.plocod}{Defalult TRUE. The function checks coherence
in ploidy and number of digits coding alleles for loci data during conversion.}
}
\description{
INTERNAL CONVERSION TOOLS FOR GENETIC DATA
}
\keyword{internal}

