% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.mantel.R
\name{int.mantel}
\alias{int.mantel}
\title{Mantel and partial Mantel tests, internal.}
\usage{
int.mantel(
  d1,
  d2,
  dc,
  nsim,
  test,
  alternative = "auto",
  plotit = FALSE,
  method = "pearson",
  na.omit = FALSE,
  ...
)
}
\arguments{
\item{d1}{distance matrix.}

\item{d2}{distance matrix.}

\item{dc}{conditional distance matrix (optional).}

\item{nsim}{Number of Monte-Carlo simulations.}

\item{test}{If test = "bootstrap", the program generates a bootstrap 
resampling and the associated confidence intervals.
If test = "permutation" a permutation test is made and the P-values 
are computed.}

\item{alternative}{The alternative hypothesis. If "auto" is selected (default) the
program determines the alternative hypothesis.
Other options are: "two.sided", "greater" and "less".}

\item{plotit}{Should be generated a plot of the simulations?}
}
\description{
Mantel and partial Mantel tests, internal.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}
