% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{int.multiplot-class}
\alias{int.multiplot-class}
\title{int.multiplot class}
\description{
int.multiplot class
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

