% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.plotCorrelog.R
\name{int.multiplot}
\alias{int.multiplot}
\title{int.multiplot method. Graphical processing of multiple correlograms}
\usage{
int.multiplot(correlog, significant = TRUE, ...)
}
\description{
int.multiplot method. Graphical processing of multiple correlograms
}
\keyword{internal}

