% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopop.1OF6.definition.R
\docType{class}
\name{int.popdata-class}
\alias{int.popdata-class}
\title{int.popdata class}
\description{
int.popdata class
}
\section{Slots}{

\describe{
\item{\code{ploidy}}{ploidy}

\item{\code{type}}{type of data ("codominant" or "dominant")}

\item{\code{NA.char}}{NA character}

\item{\code{ncod}}{number of digits coding each allele (codominant data)}

\item{\code{aggregator}}{function used to aggregate data}

\item{\code{factor_to_counts}}{Logical. Factors splitted into counts for each level?}

\item{\code{loc.fac}}{locus of each allele}

\item{\code{all.names}}{alleles names}
}}

\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{internal}
