% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{int.undimmattg}
\alias{int.undimmattg}
\title{Creates a matrix without diagonal, in row order}
\usage{
int.undimmattg(mat)
}
\arguments{
\item{mat}{Matrix.}
}
\description{
Creates a matrix without diagonal, in row order
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

