% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopop.3OF6.basic.methods.R
\docType{methods}
\name{is.locked,ecopop-method}
\alias{is.locked,ecopop-method}
\alias{is.locked,ecopop}
\title{Test if rows of an ecopop object are locked}
\usage{
\S4method{is.locked}{ecopop}(object)
}
\description{
Test if rows of an ecopop object are locked
}
