% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{meta2char}
\alias{meta2char}
\title{Metachacter to character}
\usage{
meta2char(X)
}
\arguments{
\item{X}{character string}
}
\description{
Metachacter to character
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

