% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{misc.2symmetric}
\alias{misc.2symmetric}
\title{Conversion of a non symmetric binary matrix into symmetric.}
\usage{
misc.2symmetric(mat)
}
\arguments{
\item{mat}{Matrix.}
}
\description{
Conversion of a non symmetric binary matrix into symmetric.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

