% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{misc.undimmattg}
\alias{misc.undimmattg}
\title{Creates a matrix without diagonal, in row order}
\usage{
misc.undimmattg(mat)
}
\arguments{
\item{mat}{Matrix.}
}
\description{
Creates a matrix without diagonal, in row order
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{internal}
