% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{my_ecopop}
\alias{my_ecopop}
\title{my_ecopop}
\usage{
data(eco.test)
my_ecopop
}
\description{
ecopop object generated with the object eco
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{data}
