% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{P}
\alias{P}
\title{P}
\usage{
data(eco4)
P
}
\description{
Factor with simulated phenotypic data of 173 individuals.
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{data}
