% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phenotype.R
\docType{data}
\name{phenotype}
\alias{phenotype}
\title{phenotype}
\usage{
data(eco.test)
phenotype
}
\description{
Data frame with simulated morphometric
data of 173 individuals.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{data}

