% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.methods.R
\docType{methods}
\name{plot,eco.gsa,ANY-method}
\alias{plot,eco.gsa,ANY-method}
\alias{plot,eco.gsa-method}
\title{Plot for eco.gsa objects}
\usage{
\S4method{plot}{eco.gsa,ANY}(x)
}
\arguments{
\item{x}{eco.gsa object}
}
\description{
Plot method for an eco.gsa objects object.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}
