% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.methods.R
\docType{methods}
\name{plot,eco.weight,ANY-method}
\alias{plot,eco.weight,ANY-method}
\alias{plot,eco.weight-method}
\title{Plot for a connection network}
\usage{
\S4method{plot}{eco.weight,ANY}(x)
}
\arguments{
\item{con}{connection network}
}
\description{
Plot method for an eco.weight object. This function
makes a graph for the original coordinates and transformed as ranks.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

