% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/other.classes.R
\docType{class}
\name{RelDist-class}
\alias{RelDist-class}
\title{RelDist-class}
\description{
RelDist-class
}
\section{Slots}{

\describe{
\item{\code{OUT}}{output from eco.autocor analysis}

\item{\code{INTERVAL}}{interval}

\item{\code{MAX}}{max distance}

\item{\code{TYPE}}{type}

\item{\code{NAMES}}{names}

\item{\code{METHOD}}{method}
}}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

