% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ecogen.4OF5.brackets.R
\docType{methods}
\name{[[,ecogen,numeric,missing-method}
\alias{[[,ecogen,numeric,missing-method}
\title{[[}
\usage{
\S4method{[[}{ecogen,numeric,missing}(x, i, j)
}
\description{
[[
}
\keyword{internal}

