% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{tab}
\alias{tab}
\title{tab}
\usage{
data(tab)
tab
}
\description{
Data frame with information of bands 3 and 4 for two
dates, corresponding to real Landsat 5 images and used in this
package as pedagogic material complementing simulated data.
Date and sun elevation data were extracted from the header
provided with the image in \url{http://glovis.usgs.gov/}.
The starting haze values (SHV) were estimated checking the
profiles of the bands.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{data}

