% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{table.sokal}
\alias{table.sokal}
\title{table.sokal}
\usage{
data(sokal1986)
table.sokal
}
\description{
Allelic frequency table from 50 villages, analyzed in Sokal et al. (1986).
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{data}

