% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{XY}
\alias{XY}
\title{XY}
\usage{
data(eco4)
XY
}
\description{
Factor with simulated coordinates of 173 individuals.
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{data}
