% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.aggregated_df}
\alias{aue.aggregated_df}
\title{Generate aggregated dataframe}
\usage{
aue.aggregated_df(X, hier, fun, factor_to_counts = FALSE, ...)
}
\arguments{
\item{X}{data frame}

\item{hier}{hierarchy}

\item{fun}{function}

\item{factor_to_counts}{split factor into counts for each level?}

\item{...}{additional parameters passed to fun}
}
\description{
Generate aggregated dataframe
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
