% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.df2image}
\alias{aue.df2image}
\title{Transforming a data frame into a raster}
\usage{
aue.df2image(x, origin = "upperleft")
}
\arguments{
\item{x}{output of aue.image2df}
}
\description{
This is the inverse function of aue.image2df
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

