% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.filter}
\alias{aue.filter}
\title{Local filter}
\usage{
aue.filter(mat, r, fun)
}
\arguments{
\item{mat}{Input raster matrix.}

\item{r}{Radius of the filter in pixels.}

\item{fun}{Function to apply.}
}
\description{
This program applies a function defined by the user,
into a circle of radius r around each pixel of a raster, assigning
the result to the focal pixel.
}
\examples{
\dontrun{

ras <- matrix(eco$P[,1],15,15)
image(ras)
ras.mean <- aue.filter(ras, 3, mean)
image(ras.mean)
}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

