% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.rescale}
\alias{aue.rescale}
\title{Scaling a data frame or matrix to [0, 1] or [-1, 1] ranges}
\usage{
aue.rescale(dfm, method = c("zero.one", "one.one"))
}
\arguments{
\item{dfm}{Data frame, matrix or vector to scale.}

\item{method}{Scaling method: "zero.one" for scaling to [0, 1], "one-one" for scaling to [-1,1].}
}
\description{
This program scales each column of a data frame or a matrix 
to [0,1] range, computing ((X)\emph{ij} - (Xmin)\emph{i}) / range((X)\emph{i}) 
for each individual \emph{j} of the variable \emph{i} or to [-1,1] range 
computing  2 *((X)\emph{ij} - (Xmin)\emph{i}) / range((X)\emph{i}) -1
}
\examples{

\dontrun{

data(eco.test)
require(adegenet)
pc <- dudi.pca(eco[["P"]], scannf = FALSE, nf = 3)
pc <- pc$li
pc <- aue.rescale(pc)
plot(eco[["XY"]][, 1], eco[["XY"]][, 2], col = rgb(pc), pch = 16, 
cex = 1.5, xlab ="X", ylab= "Y")

}

}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

