% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{aue.square}
\alias{aue.square}
\title{Solid square}
\usage{
aue.square(mat, d, x0, y0)
}
\arguments{
\item{mat}{Input raster matrix.}

\item{x0}{Square center in the x direction.}

\item{y0}{Square center in the y direction.}

\item{r}{Radius of the square in pixels.}
}
\description{
Solid square
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

