% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coordinates.R
\docType{data}
\name{coordinates}
\alias{coordinates}
\title{Coordinates}
\usage{
data(eco.test)
coordinates
}
\description{
Data frame with simulated UTM coordinates of
173 individuals.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{data}

