% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eco.2genepop.R
\name{eco.2genepop}
\alias{eco.2genepop}
\title{Exporting an ecogen genetic data frame into Genepop format}
\usage{
eco.2genepop(eco, x, ndig)
}
\arguments{
\item{eco}{ecogen object.}

\item{x}{The name of the S slot column with the groups for the analysis.}

\item{ndig}{number of digits coding each allele (1: x, 2: xx or
3:xxx) when there is more than one allele per individual locus.}
}
\value{
The exported file \emph{infile.genepop.txt} in the workspace.
}
\description{
Exporting an ecogen genetic data frame into Genepop format
}
\examples{
data(eco.test)
\dontrun{

eco.2genepop(eco, "structure", 1)
#the output file "infile.genepop.txt" is in the workspace

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

