% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.2gstudio.R
\name{eco.2gstudio}
\alias{eco.2gstudio}
\title{Converting a diploid ecogen genetic data frame into a gstudio object}
\usage{
eco.2gstudio(eco, type = "separated", ...)
}
\arguments{
\item{eco}{Object of class "ecogen".}

\item{type}{The type of data passed to gstudio locus. 
Default is "separated" (data as microsatellites or individual haplotypes);
"aflp" for presence - absence data.}

\item{...}{Further arguments passed to \code{\link[adegenet]{df2genind}}.}
}
\description{
This function converts the genetic 
data of an ecogen object in a gstudio data frame.
}
\examples{

\dontrun{

data(eco.test)
gsteco <- eco.2gstudio(eco, "separated")
gsteco

}

}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

