% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eco.append.R
\name{eco.append}
\alias{eco.append}
\title{Creating an updated ecogen object by adding
results to the slot OUT}
\usage{
eco.append(eco, ...)
}
\arguments{
\item{eco}{ecogen object.}

\item{...}{objects to include in eco, typed without quotations.}
}
\description{
Creating an updated ecogen object by adding
results to the slot OUT
}
\examples{
\dontrun{

data(eco.test)
variog <- eco.variogram(eco, eco$P$P1)
eco <- eco.append(eco, variog)
we.are.integers <- c(1:10)
we.are.characters <- c("John Coltrane", "Charlie Parker")

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

