% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eco.clear.R
\name{eco.clear}
\alias{eco.clear}
\title{Clearing the working environment, maintaining only the specified objects.}
\usage{
eco.clear(...)
}
\arguments{
\item{...}{Objects to retain.}
}
\description{
Clearing the working environment, maintaining only the specified objects.
}
\examples{
\dontrun{

data(eco.test)
ls()
eco.clear(eco)
ls()

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

