% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.methods.R
\docType{methods}
\name{plot,eco.listlsa,ANY-method}
\alias{plot,eco.listlsa,ANY-method}
\alias{plot,eco.listlsa-method}
\title{listplot}
\usage{
\S4method{plot}{eco.listlsa,ANY}(x)
}
\arguments{
\item{x}{list of ggplot objects}

\item{n}{number of plot in layout}

\item{nrow}{Number of rows in layout}

\item{byrow}{plot by row?}

\item{significant}{plot only significant individuals?}
}
\description{
Plot method for local spatial analysis as list
}
\seealso{
\code{\link{eco.lsa}}
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{internal}
