% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/other.classes.R
\docType{methods}
\name{show,eco.mlm-method}
\alias{show,eco.mctree-method}
\alias{show,eco.mlm-method}
\title{show eco.mlm}
\usage{
\S4method{show}{eco.mlm}(object)

\S4method{show}{eco.mctree}(object)
}
\description{
show eco.mlm

show eco.mctree
}
\keyword{internal}

