% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.nei_distance.R
\name{eco.nei_distance}
\alias{eco.nei_distance}
\title{Estimate Nei distance matrix}
\usage{
eco.nei_distance(obj, as_dist = TRUE)
}
\arguments{
\item{obj}{ecopop orgenpop  objects, or matrix/data frame with allele frequencies}

\item{as_dist}{Return a dist object or a matrix? default an object of class "dist".}
}
\description{
Estimate Nei distance matrix. NAs are avoided.
}
\examples{
\dontrun{
data(eco.test)
eco.nei_distance(my_ecopop)
}
}
\author{
Juan Vilardi
}
