% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eco.rbind.R
\name{eco.rbind}
\alias{eco.rbind}
\title{Combining the rows of two ecogen objects}
\usage{
eco.rbind(eco1, eco2, ..., check.col.names = TRUE)
}
\arguments{
\item{eco1}{Object of class "ecogen".}

\item{eco2}{Object of class "ecogen".}

\item{...}{Other "ecogen" objects to combine.}

\item{check.col.names}{Check for duplicated column names? Default TRUE.}
}
\description{
Combining the rows of two ecogen objects
}
\examples{
\dontrun{

data(eco.test)

# duplicated row names are not allowed by eco.rbind.

eco2 <- eco

# row names not duplicated for P

rownames(eco2[["P"]]) <-226:450

eco.r <- eco.rbind(eco, eco2)

eco.r

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

