% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eco.rbind.R
\name{eco.rbind}
\alias{eco.rbind}
\title{Combining the rows of two ecogen objects}
\usage{
eco.rbind(e1, e2)
}
\arguments{
\item{e1}{Object of class "ecogen".}

\item{e2}{Object of class "ecogen".}
}
\description{
Combining the rows of two ecogen objects
}
\examples{
\dontrun{

data(eco.test)
eco2<-eco.rbind(eco,eco)
}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

