% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eco.R
\docType{data}
\name{eco}
\alias{eco}
\title{Eco}
\usage{
data(eco.test)
eco
}
\description{
EcoGen object with simulated data of 173 individuals.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{data}

