% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{eco}
\alias{eco}
\title{eco}
\usage{
data(eco.test)
eco
}
\description{
ecogen object with simulated data of 225 individuals, with codominant markers
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{data}
