% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eco.slide.con.R
\name{eco.slide.con}
\alias{eco.slide.con}
\title{Sliding window for connection networks}
\usage{
eco.slide.con(x, con, fun)
}
\arguments{
\item{fun}{Function to apply in each focal pixel.}

\item{var}{Input variable or matrix}

\item{r}{half a side for square, radius for circle, diagonal length for rhombus.}

\item{slide}{number of elements between two focal pixels, for column
and row dimension}

\item{window}{window type. Default "square".}

\item{within}{should be computed the function in borders focal pixels only if
the area is within the matrix? Default TRUE.}
}
\description{
This program applies a function defined by the user, over the individuals
included in a connection network.
For a given variable, the program computes recursively a function for the individual
of the network, using all the individuals connected to each.
The function uses a connection network product of the function eco.weight
}
\examples{
\dontrun{

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

