% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eco.subset.R
\name{eco.subset}
\alias{eco.subset}
\title{Subsetting an ecogen object by group}
\usage{
eco.subset(object, fact, grp, missing = c(0, "NA", "MEAN"), ...)
}
\arguments{
\item{object}{Ecogen object.}

\item{fact}{The name of the S slot column with the groups for the analysis.}

\item{grp}{The group in the column x in which te partition will be}

\item{missing}{Missing Argument passed to \code{\link[adegenet]{df2genind}}
This can take three values as described in the latter ("0", "NA" or "MEAN").
Missing elements are treated as zeros in the default option.}

\item{...}{Further arguments passed to \code{\link[adegenet]{df2genind}}.}
}
\description{
Subsetting an ecogen object by group
}
\examples{
\dontrun{

data(eco.test)
eco<-eco.subset(eco,"structure", 1)

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

