% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{eco2}
\alias{eco2}
\title{Eco2}
\usage{
data(eco2)
eco2
}
\description{
ecogen object with simulated data of 900 individuals.
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{data}
