% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{eco3}
\alias{eco3}
\title{Eco3}
\usage{
data(eco3)
eco3
}
\description{
ecogen object with simulated data of 173 individuals.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{data}

