% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ecogen.1OF5.definition.R
\docType{class}
\name{ecogen-class}
\alias{ecogen-class}
\title{ecogen class}
\description{
ecogen class
}
\section{Slots}{

\describe{
\item{\code{XY}}{P data frame}

\item{\code{P}}{P data frame}

\item{\code{G}}{G data frame}

\item{\code{A}}{A data frame}

\item{\code{E}}{E data frame}

\item{\code{S}}{S data frame}

\item{\code{C}}{C data frame}

\item{\code{OUT}}{results}

\item{\code{INT}}{int.gendata slot}
}}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

