% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ecogen.definition.R
\docType{methods}
\name{show,ecogen-method}
\alias{$,ecogen,character-method}
\alias{$,ecogen-method}
\alias{$<-,ecogen,character,ANY-method}
\alias{$<-,ecogen-method}
\alias{[,ecogen,integer,missing,ANY-method}
\alias{[,ecogen,integer,missing-method}
\alias{is,ecogen-method}
\alias{is.ecogen}
\alias{names,ecogen-method}
\alias{show,ecogen-method}
\title{show method}
\usage{
\S4method{show}{ecogen}(object)

\S4method{$}{ecogen}(x, name)

\S4method{$}{ecogen}(x, name) <- value

\S4method{names}{ecogen}(x)

is.ecogen(x)

\S4method{[}{ecogen,integer,missing,ANY}(x, i, j, ..., drop = FALSE)
}
\description{
show method

$

$<-

names

is.ecogen

[
}
\keyword{internal}

