% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ecogen.3OF5.basic.methods.R, R/ecogen.4OF5.brackets.R
\docType{methods}
\name{names,ecogen-method}
\alias{$,ecogen,character-method}
\alias{$,ecogen-method}
\alias{$<-,ecogen,character,ANY-method}
\alias{$<-,ecogen-method}
\alias{[,ecogen,character,missing,ANY-method}
\alias{[,ecogen,integer,missing-method}
\alias{[,ecogen,numeric,missing-method}
\alias{[[,ecogen,character,missing-method}
\alias{[[<-,ecogen,character,missing-method}
\alias{as.int.list}
\alias{as.list,ecogen-method}
\alias{is,ecogen-method}
\alias{is.ecogen}
\alias{names,ecogen-method}
\alias{show,ecogen-method}
\title{names}
\usage{
\S4method{names}{ecogen}(x)

is.ecogen(x)

\S4method{as.list}{ecogen}(x)

\S4method{as.int.list}{ecogen}(x)

\S4method{show}{ecogen}(object)

\S4method{$}{ecogen}(x, name)

\S4method{$}{ecogen}(x, name) <- value

\S4method{[}{ecogen,numeric,missing}(x, i, j, ..., drop = FALSE)

\S4method{[[}{ecogen,character,missing}(x, i, j)

\S4method{[[}{ecogen,character,missing}(x, i, j, ...) <- value
}
\arguments{
\item{x}{Object of class ecogen}

\item{x}{Object of class ecogen}
}
\description{
names

is.ecogen

as.list

as.int.list

show method

$

$<-

[

[[

[[<- -----------------------------------------------------------------------#
}
\keyword{internal}

