% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecogen.6OF6.converters.R
\name{ecogen2geneland}
\alias{ecogen2geneland}
\title{Creating input data for Geneland with an ecogen object}
\usage{
ecogen2geneland(eco, dir = "", ncod = NULL, ploidy = 2,
  to_numeric = FALSE, nout = 3, recode = c("column", "all"), ...)
}
\arguments{
\item{eco}{Object of class "ecogen"}

\item{dir}{output path. Default = "" (current directory).}

\item{ncod}{Number of digits coding each allele
(e.g., 1: x, 2: xx, 3: xxx, etc.).}

\item{ploidy}{Ploidy of the data.}

\item{to_numeric}{Recode the genetic data into numeric format? If TRUE, 
the functions performs the correction via \code{\link{eco.format}}.
Additional formatting parameters can be passed to this function.}

\item{nout}{Number of digits in the output when to_numeric = TRUE.}

\item{recode}{if recode = TRUE,recoding should be performed individually (e.g., microsatellite data)
for each column, or overall (e.g., SNPs). Default: "column".}

\item{...}{Additional parameters passed to \code{\link{eco.format} when to_numeric = TRUE}}
}
\value{
XY.txt Matrix with coordinates.

NAMES.txt Matrix with row names.

P.txt Matrix with phenotypic data.

G.txt Matrix with genotypic data.
}
\description{
This function creates four data frames (XY.txt, NAMES.txt, P.txt, G.txt) 
in the indicated directory (default: working directory), which can be loadedin Geneland.
}
\examples{

\dontrun{

data(eco.test)
ecogen2geneland(eco, dir = "", ncod=1)

}

}
\author{
Leandro Roser \email{learoser@gmail.com}
}
