% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecogen.6OF6.converters.R
\name{ecogen2genepop}
\alias{ecogen2genepop}
\title{Exporting an ecogen genetic data frame into Genepop format}
\usage{
ecogen2genepop(eco, dir = "", outName = "infile.genepop.txt", grp = NULL,
  nout = 3, sep = "", recode = "none")
}
\arguments{
\item{eco}{Object of class "ecogen".}

\item{dir}{output path. Default = "" (current directory).}

\item{outName}{The name of the output file.}

\item{grp}{The name of the S slot column with groups in which the sample
must be divided (e.g., populations). If groups are not given (grp = NULL),
all individuals will be assigned to a single one.}

\item{nout}{Number of digits in the output file}

\item{sep}{Character separating alleles.}

\item{recode}{if recode = TRUE, should recoding be performed individually (e.g., microsatellite data)
for each column, or overall (e.g., SNPs). Default: "column".}

\item{...}{Additional parameters passed to \code{\link{eco.format}}}
}
\value{
A Genepop file in the working directory.
}
\description{
This function converts the genetic 
data of an ecogen object into a Genepop input file.
}
\examples{

\dontrun{

data(eco.test)
ecogen2genepop(eco, dir = "", outName = "infile.genepop.txt", grp = "pop")
# an output file "infile.genepop.txt" is generated in the working directory

}

}
\author{
Leandro Roser \email{learoser@gmail.com}
}
