% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ecogen.definition.R
\name{genind2ecogen}
\alias{ecogen2genind}
\alias{genind2ecogen}
\title{Conversion between genind and ecogen, and vice versa}
\usage{
genind2ecogen(gen)

ecogen2genind(eco)
}
\arguments{
\item{gen}{genind object}

\item{eco}{ecogen object}
}
\description{
Conversion between genind and ecogen, and vice versa


}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}

