% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecopop.6OF6.converters.R
\name{ecopop2genpop}
\alias{ecopop2genpop}
\title{Conversion form ecopop to genpop and genpop to ecopop}
\usage{
ecopop2genpop(from)
}
\arguments{
\item{from}{Object of class "ecopop" / "genpop"}
}
\description{
These functions export from ecopop to genpop and viceversa
}
\examples{

\dontrun{
data(eco.test)
my_ecopop <- ecogen2ecopop(eco, hier = "pop")
ecpop2genpop(my_ecopop)
}

}
\author{
Leandro Roser \email{learoser@gmail.com}
}
