% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{genotype}
\alias{genotype}
\title{genotype}
\usage{
data(eco.test)
genotype
}
\description{
Data frame with simulated microsatellite 
data of 225 individuals.
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{data}
