% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{int.check.con}
\alias{int.check.con}
\title{Check a connection network}
\usage{
int.check.con(con)
}
\arguments{
\item{con}{Connection network.}
}
\description{
Check a connection network
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}
