% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{int.check.group}
\alias{int.check.group}
\title{Check factor name consistency in a data frame and returns the corresponding column}
\usage{
int.check.group(X, grp = NULL, dummy = TRUE, exp.l = NULL)
}
\arguments{
\item{X}{Matrix or data frame}

\item{lab}{String used as label}
}
\description{
Check factor name consistency in a data frame and returns the corresponding column
}
\keyword{internal}
