% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/control.R
\name{int.check.ncod}
\alias{int.check.ncod}
\title{Check ploidy and number of digits per allele}
\usage{
int.check.ncod(X, ploidy, ncod = NULL, sep = "", numeric.dat = FALSE)
}
\arguments{
\item{X}{Matrix to check.}

\item{ploidy}{Ploidy level in X.}

\item{ncod}{Number of digits coding each allele.}

\item{sep}{Character string separating alleles.}

\item{numeric.dat}{Numeric data checks. Default FALSE.}
}
\description{
Check ploidy and number of digits per allele
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

