% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/control.R
\name{int.check.numeric}
\alias{int.check.numeric}
\title{Check numeric format in a data frame}
\usage{
int.check.numeric(mat)
}
\arguments{
\item{x}{Matrix or data frame.}
}
\description{
Check numeric format in a data frame
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

