% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.kin.loiselle.R
\name{int.kin.loiselle}
\alias{int.kin.loiselle}
\title{obtetion of multilocus Loiselle's Fij matrix}
\usage{
int.kin.loiselle(geno, locus)
}
\arguments{
\item{geno}{genotipic matrix: eco@A for codominant, eco@G for dominant}

\item{locus}{corresponds to: as.numeric(eco@INT@loc.fac)}
}
\description{
obtetion of multilocus Loiselle's Fij matrix
}
\keyword{internal}
