% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.methods.R
\docType{methods}
\name{int.multiplot}
\alias{int.multiplot}
\alias{plot,int.multiplot}
\alias{plot,int.multiplot,ANY-method}
\title{int.multiplot method. Graphical processing of multiple correlograms}
\usage{
int.multiplot(correlog, significant = TRUE, plotit = TRUE, ...)

\S4method{plot}{int.multiplot,ANY}(x, xlabel = NULL, ylabel = NULL,
  title = NULL, background = c("grey", "white"), meanplot = TRUE,
  multiplot = TRUE)
}
\description{
int.multiplot method. Graphical processing of multiple correlograms

plot int.multiplot
}
\keyword{internal}

