% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.random.test.R
\name{int.random.test}
\alias{int.random.test}
\title{random test}
\usage{
int.random.test(repsim = NULL, obs, nsim, test = c("permutation",
  "bootstrap"), alternative = c("auto", "two.sided", "greater", "less"),
  adjust = "none", alpha = 0.05)
}
\description{
random test
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{internal}

