% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecogen.3OF6.basic.methods.R
\docType{methods}
\name{is.locked,ecogen-method}
\alias{is.locked,ecogen-method}
\alias{is.locked,ecogen}
\title{Test if rows of an ecogen object are locked}
\usage{
\S4method{is.locked}{ecogen}(object)
}
\arguments{
\item{object}{ecogen object}
}
\description{
Test if rows of an ecogen object are locked
}
\examples{
\dontrun{
data(eco.test)
is.locked(eco) 
eco2 <- eco.unlock(eco)
is.locked(eco2) 
}
}
