% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{phenotype}
\alias{phenotype}
\title{phenotype}
\usage{
data(eco.test)
phenotype
}
\description{
Data frame with simulated morphometric 
data of 225 individuals.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{data}

