% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/other.classes.R
\docType{methods}
\name{plot,RelDist,ANY-method}
\alias{plot,RelDist,ANY-method}
\alias{plot,RelDist-method}
\title{Plot for RelDist objects.}
\usage{
\S4method{plot}{RelDist,ANY}(x, var)
}
\arguments{
\item{x}{RelDist object.}

\item{var}{Variable to plot (numeric, see examples).}
}
\description{
Plot for RelDist objects.
}
\examples{
\dontrun{

eco.ac <- eco.autocor(eco, int = 50, smax = 1000,
d.class = "phenotype")
plot(eco.ac, var = 6)
plot(eco.ac, var = 3)

}
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\seealso{
\code{\link{eco.autocor}}
}

