% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/structure.R
\docType{data}
\name{structure}
\alias{structure}
\title{structure}
\usage{
data(eco.test)
structure
}
\description{
Factor with simulated groups of 173 individuals.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{data}

