% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.auxiliar.R
\docType{data}
\name{structure}
\alias{structure}
\title{structure}
\usage{
data(eco.test)
structure
}
\description{
Factor with simulated groups of 225 individuals.
}
\author{
Leandro Roser \email{leandroroser@ege.fcen.uba.ar}
}
\keyword{data}

