% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecogen.4OF6.brackets.R
\name{[[,ecogen,numeric,missing-method}
\alias{[[,ecogen,numeric,missing-method}
\title{[[}
\usage{
\S4method{[[}{ecogen,numeric,missing}(x, i, j)
}
\description{
[[
}
\author{
Leandro Roser \email{learoser@gmail.com}
}
\keyword{internal}
